+++ src/DotPulsar/RoundRobinPartitionRouter.cs
@@ -38,13 +38,13 @@ namespace DotPulsar
         /// <summary>
         /// Choose a partition in round robin routing mode
         /// </summary>
-        public int ChoosePartition(MessageMetadata? messageMetadata, int partitionsCount)
+        public int ChoosePartition(MessageMetadata? messageMetadata, int numberOfPartitions)
         {
             var keyBytes = messageMetadata?.KeyBytes;
             if (keyBytes is not null)
-                return (int) MurmurHash3.Hash32(keyBytes, 0) % partitionsCount;
+                return (int) MurmurHash3.Hash32(keyBytes, 0) % numberOfPartitions;
 
-            return Interlocked.Increment(ref _partitionIndex) % partitionsCount;
+            return Interlocked.Increment(ref _partitionIndex) % numberOfPartitions;
         }
     }
 }