+++ src/DotPulsar/SinglePartitionRouter.cs
@@ -46,14 +46,14 @@ namespace DotPulsar
         /// <summary>
         /// Choose a partition in single partition routing mode
         /// </summary>
-        public int ChoosePartition(MessageMetadata? messageMetadata, int partitionsCount)
+        public int ChoosePartition(MessageMetadata? messageMetadata, int numberOfPartitions)
         {
             var keyBytes = messageMetadata?.KeyBytes;
             if (keyBytes is not null)
-                return (int) MurmurHash3.Hash32(keyBytes, 0) % partitionsCount;
+                return (int) MurmurHash3.Hash32(keyBytes, 0) % numberOfPartitions;
             
             if (_partitionIndex == -1)
-                _partitionIndex = new Random().Next(0, partitionsCount);
+                _partitionIndex = new Random().Next(0, numberOfPartitions);
 
             return _partitionIndex;
         }